refactor blsct to fix python dlopen issue
