+++ src/blsct/external_api/blsct.h
@@ -8,9 +8,9 @@
 #include <blsct/double_public_key.h>
 #include <blsct/private_key.h>
 #include <blsct/public_key.h>
-#include <blsct/wallet/address.h>
 #include <blsct/arith/mcl/mcl.h>
 #include <blsct/arith/elements.h>
+#include <blsct/chain.h>
 #include <blsct/range_proof/setup.h>
 #include <primitives/transaction.h>
 #include <tinyformat.h>
@@ -118,13 +118,6 @@ if (name == nullptr) err(BLSCT_MEM_ALLOC_FAILED);
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-enum Chain {
-    MainNet,
-    TestNet,
-    SigNet,
-    RegTest
-};
 
 enum TxOutputType {
     Normal,
@@ -234,7 +227,6 @@ void free_amounts_ret_val(BlsctAmountsRetVal* rv); // free attrs as well
 
 // library initialization
 void init();
-bool set_chain(enum Chain chain);
 
 // point
 BlsctRetVal* gen_random_point();
@@ -393,6 +385,11 @@ const BlsctScalar* get_tx_out_range_proof_s_prime(const CTxOut* tx_out);
 const BlsctScalar* get_tx_out_range_proof_delta_prime(const CTxOut* tx_out);
 const BlsctScalar* get_tx_out_range_proof_alpha_hat(const CTxOut* tx_out);
 const BlsctScalar* get_tx_out_range_proof_tau_x(const CTxOut* tx_out);
+
+const BlsctSignature* sign_message(
+    const BlsctScalar* blsct_priv_key,
+    const char* blsct_msg
+);
 
 bool verify_msg_sig(
     const BlsctPubKey* blsct_pub_key,
